Moves drifting into a component, this lets me explictly block input after the move loop ends, so people can't move the moment they functionally move onto a new tile This is a bit underdeveloped currently, but that's a problem for another day
Cleans up some uses of move procs, fixes runtimes in metoer and curseblob code
Adds signals for stopping/starting a move loop, sending one for destroy is redundant.
Moves existing event signals from the movable being acted on to the loop itself, makes more sense this way
Makes the move handler return the created loop up the chain so we can register to it
Fixes a logic error in loop contesting code that lead to loops never actually being removed from subsystems
because they didn't know they should be.
Properly changes lifetime from a time to stop, to functionally an amount of moves to complete before stopping
Adds some new signals for pre/post loop process. This is to better tie into components.

I decided I didn't like the idea of tying all functionality to the loops themselves
The loop decides functionally how to move, components or just tied in signals can decide when/when not to move
and can modify properties of the loop

Making a new loop for things like atmos drift, something I'm interested in tackling in the future, seemed silly

+++ code/modules/meteors/meteors.dm
@@ -128,6 +128,8 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 
 /obj/effect/meteor/Moved(atom/OldLoc, Dir, Forced = FALSE)
 	. = ..()
+	if(!OldLoc)
+		return
 
 	var/dest_turf = get_turf(dest) //Dest can be deleted in the if block
 
@@ -155,12 +157,17 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 /obj/effect/meteor/proc/chase_target(atom/chasing, delay, home)
 	if(!isatom(chasing))
 		return
-	move_towards(src, chasing, delay, home, lifetime)
-	RegisterSignal(src, COMSIG_MOVELOOP_END, .proc/handle_stopping)
+	var/datum/move_loop/new_loop = move_towards(src, chasing, delay, home, lifetime)
+	if(!new_loop)
+		return
 
+	RegisterSignal(new_loop, COMSIG_PARENT_QDELETING, .proc/handle_stopping)
+
+///Deals with what happens when we stop moving, IE we die
 /obj/effect/meteor/proc/handle_stopping()
 	SIGNAL_HANDLER
-	qdel(src)
+	if(!QDELETED(src))
+		qdel(src)
 
 /obj/effect/meteor/proc/ram_turf(turf/T)
 	//first bust whatever is in the turf
